+++ CHANGELOG.md
@@ -73,6 +73,7 @@ Version 0.7.0-SNAPSHOT - Build 20160115
 - Added custom thread pool in the parallel stream execution. Created the ForkJoinStream class and enhanced the ThreadMethods class.
 - Created a new Configuration mechanism and added the ability to configure the concurrency.
 - Updated the NgramsExtractor algorithm to export more keywords.
+- Removed the transacted option of MapDB.
 
 Version 0.6.1 - Build 20160102
 ------------------------------